backend/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
celery==5.5.3
redis==6.4.0
Pillow==11.3.0
python-decouple==3.8
